Refactor configuration and utility structure
